Slight cleanup before changing the input file system from text-based to json-based

+++ simulator/source/inc/Location.h
@@ -4,7 +4,7 @@
 class Location
 {
 private:
-	float x, y, z;
+	double x, y, z;
 
 public:
 	Location() : x(NAN), y(NAN), z(NAN) {}
